requirements.txt
python-telegram-bot>=20.0
motor>=3.3.2
EbookLib>=0.18
aiohttp>=3.9.1
python-dotenv>=1.0.0

# Per importa_storico.py
pyrogram>=2.0.10
TgCrypto>=1.2.5
